backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
reportlab==4.0.4
pandas==2.1.3
plotly==5.18.0
streamlit==1.28.1
requests==2.31.0
python-dotenv==1.0.0
alembic==1.12.1
psycopg2-binary==2.9.9
email-validator==2.1.0